Significantly improve/correct rasterization of scaled fonts
Some subtle changes to rounding can have profound implications.

Add a test suite to quickly inspect a matrix of several fonts and
sizes. The drawn hardcopy contains Unicode box drawing characters
which makes it easy to spot anomalies.

Tested (known to work well) scaled fonts include:
- Liberation Mono
- Deja Vu Sans Mono
- Fira Code
- JetBrains Mono
- Source Code Pro
- Hasklig
- Free Mono

It is still pretty easy to come across fonts that do not work so well;
however, it is anticipated that the corrections herein will have an
overall positive impact even if the end result is not fully correct,
and with rendering of the above fonts fixed, this change seems worthy
of publishing as is.