Cut measured latency by zeroing the EGL swap interval
According to Pavel Fatin's typometer (with default settings),
this change produces a significant latency reduction. No other
visual impact was detected.

Measured on a terminal size of standard geometry (80x24) with the
default font (misc-fixed 9x18), all figures in milliseconds:

On a 60Hz full-hd laptop screen:
           min   max   avg   sd
  before   7.4  13.6  10.4  1.3
  after    3.9   9.1   5.9  1.0

On a 30Hz 4K screen:
           min   max   avg   sd
  before   5.7  39.2  14.0  4.0
  after    3.8   8.5   5.1  0.9

N.B.: this latter screen had TearFree enabled, but changing that
setting did not seem to affect the result.

Both computers have AMD Radeon (renoir) with GLES Mesa 22.3.6

Apparently the performance becomes more consistent across refresh
rates. This is only natural, as we have essentially shortcut the
part of latency stemming from waiting for the next video frame.

On larger terminals (more cells to update) the difference becomes
less and less pronounced.  On a full-screen terminal occupying
the entirety of a full-hd laptop screen (at a geometry of 212x59)
the impact of this commit looks like this:

           min   max   avg   sd
  before  11.4  17.8  14.6  1.1
  after   11.5  17.0  14.4  0.9

The timings diff here is probably not significant.

N.B.: according to the manual page[1] of eglSwapInterval(), the
interval is silently clamped to the minimum supported - which is why
we just pass zero and do not bother querying EGL_MIN_SWAP_INTERVAL.

[1] https://registry.khronos.org/EGL/sdk/docs/man/html/eglSwapInterval.xhtml