GPU performance optimization: use RGBA8 format instead of RGBA32F
This is not really an optimization, rather a de-pessimization, and it
is rather depressing that it took me so many years to come across.
Because, well, _of_course_ there is no reason to use 32-bit floats per
color (plus alpha) for each pixel location - 8 bits per component is
plenty! This waste blows up GPU memory usage by no less than a factor
of 4, and thus (and perhaps even more significantly) degrades memory
locality/perf resulting in longer runtimes of the GPU program. I also
suspect there is added cost associated with image format conversion.

Sticking to the Typometer benchmark, the results for this commit are
shown for a 60 Hz 4K screen (this is not a typo, the previous commit
it was a 30 Hz 4K screen, but now I have a better cable!). The case of
the 80x24 default size is no longer interesting; it is consistently
fast (reported as around 5 ms on average). So now I tested two larger
sizes, which so far exhibited rapid slowdowns with the increase in
Zutty surface.

Half-screen size (geometry 222x117):
           min    max   avg   sd
  before   14.4  34.7  17.2  1.9
  after     6.1  22.7  11.4  1.9

Full-screen on 4K (geometry 426x119):
           min    max   avg   sd
  before   22.0  42.6  24.2  2.1
  after     9.0  19.0  10.4  0.8

The most encouraging part of this result is that with this commit the
performance becomes essentially flat (beyond a certain area) as the
size goes up. I.e., refreshing the full-screen (and it's huge at 4K!)
terminal does not cost more time than refreshing only ~half of it.

Since we cannot ever expect less than half the frame period (16.6 ms
at 60 Hz) on average, these latency numbers seem reasonably optimal.

N.B.: font rendering results (image hashes) for exactly all scaled
fonts had to be updated. This is interesting, but ultimately explained
by the change in GL code, and boils down to 1-LSB rounding diffs in
intermediate brightness values (so the bitmap fonts are not affected).
There is absolutely no visible before/after difference, which was
investigated by careful A/B testing of the test result images.