Use FreeType underline metrics when rendering a scaled font
FreeType provides `underline_position' and `underline_thickness' as
part of the global font metrics which to this point have not been
used. Cells with the underline attribute have been rendered with a
fixed, one pixel thick underline on the bottommost row. This is fully
adequate for bitmap fonts (and FreeType provides no underline metrics
for these), but scaled fonts, especially at larger sizes, deserve to
be treated better.

This commit adds underlines that are rasterized according to the
fractional pixel metrics provided by FreeType. Due to the fractional
pixel rendering (i.e., anti-aliasing) that we employ, it was also
necessary to unify this with the actual glyph rendering so that the
two anti-aliased raster results may be correctly composited. This is
all handled by the compute shader.

As a natural consequence, test results of scaled font renderings are
all changed (for the better).